+++ website/docs/components/widget-empty/examples/nothing-found.tsx
@@ -8,12 +8,25 @@ import { Text } from '@semcore/ui/typography';
 import Button from '@semcore/ui/button';
 import Divider from '@semcore/ui/divider';
 
-const options = ['de', 'en', 'es', 'fr', 'it', 'ja', 'pt', 'ru', 'zh', 'ko', 'vi', 'pl', 'sv'].map(
-  (o) => ({
-    value: o,
-    children: o,
-  }),
-);
+const options = [
+  'de',
+  'en',
+  'es',
+  'fr',
+  'it',
+  'ja',
+  'pt',
+  'ru',
+  'zh',
+  'ko',
+  'vi',
+  'pl',
+  'nl',
+  'sv',
+].map((o) => ({
+  value: o,
+  children: o,
+}));
 
 export default () => {
   const [lang, setLang] = React.useState('en');