+++ semcore/button/CHANGELOG.md
@@ -2,6 +2,24 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [5.9.0] - 2023-10-10
+
+### Changed
+
+- Version preminor update due to children dependencies update (`@semcore/flex-box` [5.7.5 ~> 5.8.0]).
+
+## [5.8.5] - 2023-10-06
+
+### Fixed
+
+- added `aria-busy='true'` for loading state and `aria-disable` for disabled state.
+
+## [5.8.4] - 2023-10-03
+
+### Changed
+
+- Version prepatch update due to children dependencies update (`@semcore/utils` [4.8.0 ~> 4.8.1]).
+
 ## [5.8.3] - 2023-10-02
 
 ### Changed